requirements.txt
requests
beautifulsoup4
psycopg2-binary
python-dotenv
yfinance